+++ crates/gst/src/lib.rs
@@ -13,8 +13,6 @@
 #[cfg(feature = "gstreamer")]
 mod backend;
 #[cfg(feature = "gstreamer")]
-pub mod composite;
-#[cfg(feature = "gstreamer")]
 pub mod kms;
 #[cfg(feature = "gstreamer")]
 pub mod rtsp;